fix test

+++ stdlib/SparseArrays/test/sparsevector.jl
@@ -78,7 +78,7 @@ end
     for dims in (8, (8,))
         @test exact_equal(spzeros(dims), SparseVector(8, Int[], Float64[]))
         @test exact_equal(spzeros(Float64, dims), SparseVector(8, Int[], Float64[]))
-        @test exact_equal(spzeros(Int, Float64, dims), SparseVector(8, Int[], Float64[]))
+        @test exact_equal(spzeros(Float64, Int, dims), SparseVector(8, Int[], Float64[]))
     end
 
     @testset "from list of indices and values" begin
